% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-R6Member.R
\name{R6Member}
\alias{R6Member}
\title{\if{html}{\out{<a id="R6Member"></a>}}\if{latex}{\out{\hypertarget{R6Member}{}}}
Container for members
in \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}} definitions.}
\description{
\if{html}{\out{<a id="sec-description"></a>}}\if{latex}{\out{\hypertarget{sec-description}{}}}
\ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestriction.check}}}{\ifelse{latex}{\href{#method-R6MemberRestriction-check}{\code{R6MemberRestriction.check}}}{\link[=R6MemberRestriction]{\code{R6MemberRestriction.check}}}}
A container that stores name, value and (optional) validity restrictions of type \ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestriction}}}{\ifelse{latex}{\href{#R6MemberRestriction}{\code{R6MemberRestriction}}}{\link[=R6MemberRestriction]{\code{R6MemberRestriction}}}}
for members to be used in \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}} declarations.

A \code{R6Member} object provides
\itemize{
\item a setter that checks restrictions
\item a getter returning the value
\item helper methods for using it as member (attribute or method) in \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
}

See the sections on
\ifelse{html}{\href{#sec-R6Member-details}{details}}{\ifelse{latex}{\href{#sec-R6Member-details}{details}}{\link[=R6Member]{details}}},
\ifelse{html}{\href{#sec-R6Member-examples}{examples}}{\ifelse{latex}{\href{#sec-R6Member-examples}{examples}}{\link[=R6Member]{examples}}}
and
\ifelse{html}{\href{#method-R6Member-accessor}{methods}}{\ifelse{latex}{\href{#method-R6Member-accessor}{methods}}{\link[=R6Member]{methods}}}.
}
\section{\if{html}{\out{<a id="sec-R6Member-details">Details</a>}}\if{latex}{\out{\hypertarget{sec-R6Member-details}{Details}}}}{

\code{R6Member} objects
provide a formalized interface for setting up members of
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
and provide methods for automatically generating the corresponding
\ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} arguments
\ifelse{html}{\link[R6:R6Class]{\code{public}}}{\ifelse{latex}{\code{public}}{\link[R6:R6Class]{\code{public}}}}, \ifelse{html}{\link[R6:R6Class]{\code{private}}}{\ifelse{latex}{\code{private}}{\link[R6:R6Class]{\code{private}}}} and \ifelse{html}{\link[R6:R6Class]{\code{active}}}{\ifelse{latex}{\code{active}}{\link[R6:R6Class]{\code{active}}}}.
\subsection{Validation rules for \ifelse{html}{\href{#active-R6Member-value}{\code{$value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{$value}}}{\link[=R6Member]{\code{$value}}}} \if{html}{\out{<a id="sec-details-restricted"></a>}}\if{latex}{\out{\hypertarget{sec-details-restricted}{}}}}{

Validation rules for the
\ifelse{html}{\href{#active-R6Member-value}{\code{$value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{$value}}}{\link[=R6Member]{\code{$value}}}}
field
are specified in the
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\link[=R6Member]{\code{$restrictions}}}} field.

\ifelse{html}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\link[=R6Member]{\code{$restrictions}}}} contains a list of zero or more
\ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}{\ifelse{latex}{\href{#R6MemberRestriction}{\code{R6MemberRestrictions}}}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}}
describing the validation rules for the \ifelse{html}{\href{#active-R6Member-value}{\code{$value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{$value}}}{\link[=R6Member]{\code{$value}}}} field.
Any new \ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}} is passed to the
\ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestriction$check()}}}{\ifelse{latex}{\href{#method-R6MemberRestriction-check}{\code{R6MemberRestriction$check()}}}{\link[=R6MemberRestriction]{\code{R6MemberRestriction$check()}}}} method of all elements in \ifelse{html}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{$restrictions}}}{\link[=R6Member]{\code{$restrictions}}}}
before \ifelse{html}{\href{#active-R6Member-value}{\code{R6Member$value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{R6Member$value}}}{\link[=R6Member]{\code{R6Member$value}}}} is changed.
This ensures all validation criteria are met.
}

\subsection{Combining multiple \code{R6Members} and declaring \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}} \if{html}{\out{<a id="sec-details-multiple"></a>}}\if{latex}{\out{\hypertarget{sec-details-multiple}{}}}}{

For declaring \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}},
\code{R6Members} are most conveniently combined in a \ifelse{html}{\link[=R6MemberList]{\code{R6MemberList}}}{\ifelse{latex}{\href{#R6MemberList}{\code{R6MemberList}}}{\link[=R6MemberList]{\code{R6MemberList}}}}.
}
}

\section{\if{html}{\out{<a id="sec-R6Member-examples">Examples</a>}}\if{latex}{\out{\hypertarget{sec-R6Member-examples}{Examples}}}}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# ===================== > Create R6MemberRestrictions < ====================== #

# Restriction that checks if length of value equals 2
length_restriction <-
    R6MemberRestriction$new(
        id = "Length check",
        restriction = function(x) \{
            if (length(x) != 2) \{
                stop("Length should be 2!")
            \}
            return(TRUE)
        \}
    )

# Restriction that checks if type of value is "character"
type_restriction <-
    R6MemberRestriction$new(
        id = "Type check",
        restriction = function(x) \{
            if (typeof(x) != "character") \{
                stop("type should be character!")
            \}
            return(TRUE)
        \}
    )
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# =========================== > Create R6Members < =========================== #

# R6Member object without restriction(s):
#   may take any value
R6Member_any <-
   R6Member$new(
       name = "example_member_any",
       value = 1:5
   )

# R6Member object with restriction(s):
#   all restriction(s) check(s) must return TRUE
R6Member_two_string <-
   R6Member$new(
       name = "example_member_char_len2",
       value = c("foo", "bar"),
       restrictions = list(
               "is length 2" = length_restriction,
               "is character" = type_restriction
       )
   )
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Initial value would violate restrictions
try(\{
R6Member_invalid <-
   R6Member$new(
       name = "example_member",
       value = 1:5,
       restrictions = list(
               "is length 2" = length_restriction,
               "is character" = type_restriction
       )
   )
\})
#> Error in private$.err_violated_restrictions(value = value, restrictions = restrictions,  : 
#>   
Invalid attempt to set
#>    `example_member$value <-
#>        1:5`
#> Assignment does not meet the following restrictions:
#>    - Length check:
#>       Length should be 2!
#>    - Type check:
#>       type should be character!
}\if{html}{\out{</div>}}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{is_method}}{\if{html}{\out{<a id="active-R6Member-is_method"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-is_method}{}}}Whether the
\code{R6Member}
should be treated as a method (function) member.
\if{html}{\out{<br>}}
Typically determined automatically by checking if
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
is a
\ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}}, but can be assigned as well.
When \code{TRUE}, the member will be declared as a method in the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
definition.
\if{html}{\out{<br>}}\emph{This field automatically resets every time  \ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}  is set or changed, based on whether the new value is a function.}\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{is_private}}{\if{html}{\out{<a id="active-R6Member-is_private"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-is_private}{}}}Determines whether the
\code{R6Member}
becomes a private member of the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}.
\if{html}{\out{<br>}}
If \code{TRUE}, the member is only accessible within the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
object (e.g., in methods) and not exposed in the public interface.
In this case,
\ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}}
and
\ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}}
will return
\ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}}. If \code{FALSE} (the default), the member is public and can be accessed externally.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{FALSE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{locked}}{\if{html}{\out{<a id="active-R6Member-locked"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-locked}{}}}Determines whether the
\code{R6Member}
object's
fields
are locked (\code{TRUE}, the default) or writable (\code{FALSE}).
When locked, most fields cannot be modified after initialization.
This applies to all fields except itself, but can be overwritten for the
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
by setting
\ifelse{html}{\href{#active-R6Member-value_locked}{\code{value_locked}}}{\ifelse{latex}{\href{#active-R6Member-value_locked}{\code{value_locked}}}{\link[=R6Member]{\code{value_locked}}}}
Locking provides immutability guarantees and prevents accidental modifications to the member definition.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{TRUE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
<br>
}}}

\item{\code{name}}{\if{html}{\out{<a id="active-R6Member-name"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-name}{}}}Name of the
\code{R6Member}
object.
This name serves as the identifier for the member in the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
definition.
\if{html}{\out{<br>}}\emph{Must be a valid R variable name and not be empty.}
The name is used to generate member declarations in public, private, and active binding contexts.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
<br>
}}}

\item{\code{private_name}}{\if{html}{\out{<a id="active-R6Member-private_name"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-private_name}{}}}The private internal name of the
\code{R6Member}
object.
This is the name used to store the member in the private environment of the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}.
It is constructed from the public name with a prefix (default: \code{.}) and optional suffix.
\if{html}{\out{<br>}}\emph{This is a read-only field and is automatically computed from the public name.}\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
}
\if{html}{\out{
<br>
}}}

\item{\code{restrictions}}{\if{html}{\out{<a id="active-R6Member-restrictions"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-restrictions}{}}}A
\ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}}
of validity restrictions
(\ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}{\ifelse{latex}{\href{#R6MemberRestriction}{\code{R6MemberRestrictions}}}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}})
the
\code{R6Member}'\code{s}
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
must fulfill to be accepted.
When the value is set or modified, it is checked against all restrictions in sequence.
If any restriction fails, an error is raised and the value is not changed.
An empty list means no restrictions are applied.
See
\ifelse{html}{\href{#sec-R6Member-details}{details}}{\ifelse{latex}{\href{#sec-R6Member-details}{details}}{\link[=R6Member]{details}}}
and
\ifelse{html}{\href{#sec-R6Member-examples}{examples}}{\ifelse{latex}{\href{#sec-R6Member-examples}{examples}}{\link[=R6Member]{examples}}}
for more information on validation.\tabular{ll}{
    \tab  \cr
   Type:   \tab list          \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{value}}{\if{html}{\out{<a id="active-R6Member-value"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-value}{}}}Value of the
\code{R6Member}
object.
Can be any R object (vector, list, function, etc.).
If a function is assigned, the
\ifelse{html}{\href{#active-R6Member-is_method}{\code{is_method}}}{\ifelse{latex}{\href{#active-R6Member-is_method}{\code{is_method}}}{\link[=R6Member]{\code{is_method}}}}
field is automatically set to
\code{TRUE}.
It must meet all
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\link[=R6Member]{\code{restrictions}}}}
unless it is
\ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}}.
Any attempt to modify the value is first validated against all
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\link[=R6Member]{\code{restrictions}}}}
before the assignment is accepted.\tabular{ll}{
    \tab  \cr
   Type:   \tab any           \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{value_locked}}{\if{html}{\out{<a id="active-R6Member-value_locked"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-value_locked}{}}}Determines whether the
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
is locked (\code{TRUE}) or writable (\code{FALSE}).\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{FALSE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6Member-accessor}{\code{R6Member$accessor()}}
\item \href{#method-R6Member-new}{\code{R6Member$new()}}
\item \href{#method-R6Member-print}{\code{R6Member$print()}}
\item \href{#method-R6Member-declare_active_binding}{\code{R6Member$declare_active_binding()}}
\item \href{#method-R6Member-declare_method}{\code{R6Member$declare_method()}}
\item \href{#method-R6Member-declare_initialization}{\code{R6Member$declare_initialization()}}
\item \href{#method-R6Member-clone}{\code{R6Member$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-accessor"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-accessor}{}}}
\subsection{Method \code{accessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$accessor(field_value, field_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_value}}{\if{html}{\out{<a id="field_value"></a>}}\if{latex}{\out{\hypertarget{field_value}{}}}Optional new value for field \verb{<field_name>}.
If omitted the method returns the current field contents.
If provided the \code{field_value} is validated and assigned to the field.\tabular{ll}{
    \tab  \cr
   Type:   \tab See \ifelse{html}{\href{#active-R6Member-is_method}{active bindings}}{\ifelse{latex}{\href{#active-R6Member-is_method}{active bindings}}{\link[=R6Member]{active bindings}}} for the valid type of each field. \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{field_name}}{\if{html}{\out{<a id="field_name"></a>}}\if{latex}{\out{\hypertarget{field_name}{}}}The name of the internal field to be accessed.
\if{html}{\out{<br>}}
See
\ifelse{html}{\href{#active-R6Member-is_method}{active bindings}}{\ifelse{latex}{\href{#active-R6Member-is_method}{active bindings}}{\link[=R6Member]{active bindings}}}
for a list of all fields in
\code{R6Member}
objects.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Access or modify one of the internal fields of \code{R6Member}s
}

\subsection{Returns}{
\tabular{ll}{
    \tab  \cr
   If \code{field_value} is \strong{not} provided (getter case): \tab the field's current value -- see \ifelse{html}{\href{#active-R6Member-is_method}{active bindings}}{\ifelse{latex}{\href{#active-R6Member-is_method}{active bindings}}{\link[=R6Member]{active bindings}}} for the fields' types. \cr
   If \code{field_value} is provided (setter case):  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}}                     \cr
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$new(
  name,
  value = NULL,
  restrictions = NULL,
  locked = TRUE,
  value_locked = FALSE,
  is_private = FALSE,
  is_method = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\if{html}{\out{<a id="active-R6Member-name"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-name}{}}}Name of the
\code{R6Member}
object.
This name serves as the identifier for the member in the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
definition.
\if{html}{\out{<br>}}\emph{Must be a valid R variable name and not be empty.}
The name is used to generate member declarations in public, private, and active binding contexts.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
<br>
}}}

\item{\code{value}}{\if{html}{\out{<a id="active-R6Member-value"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-value}{}}}Value of the
\code{R6Member}
object.
Can be any R object (vector, list, function, etc.).
If a function is assigned, the
\ifelse{html}{\href{#active-R6Member-is_method}{\code{is_method}}}{\ifelse{latex}{\href{#active-R6Member-is_method}{\code{is_method}}}{\link[=R6Member]{\code{is_method}}}}
field is automatically set to
\code{TRUE}.
It must meet all
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\link[=R6Member]{\code{restrictions}}}}
unless it is
\ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}}.
Any attempt to modify the value is first validated against all
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\link[=R6Member]{\code{restrictions}}}}
before the assignment is accepted.\tabular{ll}{
    \tab  \cr
   Type:   \tab any           \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{restrictions}}{\if{html}{\out{<a id="active-R6Member-restrictions"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-restrictions}{}}}A
\ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}}
of validity restrictions
(\ifelse{html}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}{\ifelse{latex}{\href{#R6MemberRestriction}{\code{R6MemberRestrictions}}}{\link[=R6MemberRestriction]{\code{R6MemberRestrictions}}}})
the
\code{R6Member}'\code{s}
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
must fulfill to be accepted.
When the value is set or modified, it is checked against all restrictions in sequence.
If any restriction fails, an error is raised and the value is not changed.
An empty list means no restrictions are applied.
See
\ifelse{html}{\href{#sec-R6Member-details}{details}}{\ifelse{latex}{\href{#sec-R6Member-details}{details}}{\link[=R6Member]{details}}}
and
\ifelse{html}{\href{#sec-R6Member-examples}{examples}}{\ifelse{latex}{\href{#sec-R6Member-examples}{examples}}{\link[=R6Member]{examples}}}
for more information on validation.\tabular{ll}{
    \tab  \cr
   Type:   \tab list          \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{locked}}{\if{html}{\out{<a id="active-R6Member-locked"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-locked}{}}}Determines whether the
\code{R6Member}
object's
fields
are locked (\code{TRUE}, the default) or writable (\code{FALSE}).
When locked, most fields cannot be modified after initialization.
This applies to all fields except itself, but can be overwritten for the
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
by setting
\ifelse{html}{\href{#active-R6Member-value_locked}{\code{value_locked}}}{\ifelse{latex}{\href{#active-R6Member-value_locked}{\code{value_locked}}}{\link[=R6Member]{\code{value_locked}}}}
Locking provides immutability guarantees and prevents accidental modifications to the member definition.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{TRUE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
<br>
}}}

\item{\code{value_locked}}{\if{html}{\out{<a id="active-R6Member-value_locked"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-value_locked}{}}}Determines whether the
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
is locked (\code{TRUE}) or writable (\code{FALSE}).\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{FALSE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{is_private}}{\if{html}{\out{<a id="active-R6Member-is_private"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-is_private}{}}}Determines whether the
\code{R6Member}
becomes a private member of the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}.
\if{html}{\out{<br>}}
If \code{TRUE}, the member is only accessible within the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
object (e.g., in methods) and not exposed in the public interface.
In this case,
\ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}}
and
\ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}}
will return
\ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}}. If \code{FALSE} (the default), the member is public and can be accessed externally.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{FALSE}         \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{is_method}}{\if{html}{\out{<a id="active-R6Member-is_method"></a>}}\if{latex}{\out{\hypertarget{active-R6Member-is_method}{}}}Whether the
\code{R6Member}
should be treated as a method (function) member.
\if{html}{\out{<br>}}
Typically determined automatically by checking if
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
is a
\ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}}, but can be assigned as well.
When \code{TRUE}, the member will be declared as a method in the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
definition.
\if{html}{\out{<br>}}\emph{This field automatically resets every time  \ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}  is set or changed, based on whether the new value is a function.}\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \cr
   Affected methods: \tab \ifelse{html}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\ifelse{latex}{\href{#method-R6Member-declare_active_binding}{\code{declare_active_binding}}}{\link[=R6Member]{\code{declare_active_binding}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\ifelse{latex}{\href{#method-R6Member-declare_method}{\code{declare_method}}}{\link[=R6Member]{\code{declare_method}}}} \cr
      \tab \ifelse{html}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\ifelse{latex}{\href{#method-R6Member-declare_initialization}{\code{declare_initialization}}}{\link[=R6Member]{\code{declare_initialization}}}} \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Construct a new \code{R6Member} object for creating and initializing a new \code{R6Member} object.
The constructor validates all provided arguments against the specified restrictions
and locks the object after initialization.
}

\subsection{Returns}{
\code{R6Member} object with all specified fields initialized and locked
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-print"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$print(value_level = TRUE, restriction_level = 0, print_class = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value_level}}{\if{html}{\out{<a id="value_level"></a>}}\if{latex}{\out{\hypertarget{value_level}{}}}The amount of information about the
\ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}}
to be printed.
One of:\tabular{rcl}{
    \tab  \tab  \cr
   \code{value_level == FALSE} or \code{value_level <= 0} \tab  \tab print nothing        \cr
   \code{0 < value_level} \tab  \tab print \ifelse{html}{\href{#active-R6Member-value}{\code{value}'\code{s}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}'\code{s}}}{\link[=R6Member]{\code{value}'\code{s}}}} first \code{value_level} characters \cr
   \code{value_level == TRUE} \tab  \tab print \ifelse{html}{\href{#active-R6Member-value}{\code{value}}}{\ifelse{latex}{\href{#active-R6Member-value}{\code{value}}}{\link[=R6Member]{\code{value}}}} entirely \cr
}
\tabular{ll}{
    \tab  \cr
   Type:   \tab logical or integer        \cr
   Default:  \tab \code{TRUE}         \cr
}
\if{html}{\out{
<br>
When set to a positive number, this controls the maximum width for displaying the value.
If `TRUE`, displays the full value. If `FALSE` or `0`, the value is not displayed.
}}}

\item{\code{restriction_level}}{\if{html}{\out{<a id="restriction_level"></a>}}\if{latex}{\out{\hypertarget{restriction_level}{}}}The amount of information about
each of
the
\ifelse{html}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\ifelse{latex}{\href{#active-R6Member-restrictions}{\code{restrictions}}}{\link[=R6Member]{\code{restrictions}}}}
to be printed.
One of:\tabular{rcl}{
    \tab  \tab  \cr
     \code{0} \tab  \tab print nothing        \cr
     \code{1} \tab  \tab print \ifelse{html}{\link[=R6MemberRestriction]{\code{ids}}}{\ifelse{latex}{\href{#active-R6MemberRestriction-id}{\code{ids}}}{\link[=R6MemberRestriction]{\code{ids}}}} \cr
     \code{2} \tab  \tab print \ifelse{html}{\link[=R6MemberRestriction]{\code{ids}}}{\ifelse{latex}{\href{#active-R6MemberRestriction-id}{\code{ids}}}{\link[=R6MemberRestriction]{\code{ids}}}} and \ifelse{html}{\link[=R6MemberRestriction]{\code{restriction functions}}}{\ifelse{latex}{\href{#active-R6MemberRestriction-restriction}{\code{restriction functions}}}{\link[=R6MemberRestriction]{\code{restriction functions}}}} \cr
}
\tabular{ll}{
    \tab  \cr
   Type:   \tab integer          \cr
   Default:  \tab \code{0}          \cr
}
\if{html}{\out{
<br>
Controls the detail level when printing restrictions.
Higher values show more detailed restriction information.
`0` means restrictions are not printed.
}}}

\item{\code{print_class}}{\if{html}{\out{<a id="print_class"></a>}}\if{latex}{\out{\hypertarget{print_class}{}}}Whether the class name (\code{R6Member}) should be included when printing.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{TRUE}         \cr
}
\if{html}{\out{
<br>
}}


When \code{TRUE}, the class name is prepended to the member name in the output.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Print an \code{R6Member} object to the console.
Displays the member's name, optionally its value, and optionally its restrictions.
Useful for inspecting member definitions and their validation rules.
}

\subsection{Returns}{
\ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} (invisibly). Output is printed to the console.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-declare_active_binding"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-declare_active_binding}{}}}
\subsection{Method \code{declare_active_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$declare_active_binding()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Create \ifelse{html}{\link[R6:R6Class]{\code{active} binding}}{\ifelse{latex}{\code{active} binding}{\link[R6:R6Class]{\code{active} binding}}} \ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}}  for \code{R6Member} object.
\if{html}{\out{<br>}}\emph{To be used when declaring \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}},
see also
the
\ifelse{html}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{R6MemberList$R6Class()}}}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}}
method
and
\code{?R6Class}.
}

\subsection{Returns}{
\ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}}
to be used in the
\ifelse{html}{\link[R6:R6Class]{\code{active}}}{\ifelse{latex}{\code{active}}{\link[R6:R6Class]{\code{active}}}}
argument of
\ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-declare_method"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-declare_method}{}}}
\subsection{Method \code{declare_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$declare_method()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Create  \ifelse{html}{\link[R6:R6Class]{\code{method function}}}{\ifelse{latex}{\code{method function}}{\link[R6:R6Class]{\code{method function}}}} for \code{R6Member} object.
\if{html}{\out{<br>}}\emph{To be used when declaring \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}},
see also
the
\ifelse{html}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{R6MemberList$R6Class()}}}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}}
method
and
\code{?R6Class}.
}

\subsection{Returns}{
\ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}}
to be used in the
\ifelse{html}{\link[R6:R6Class]{\code{method}}}{\ifelse{latex}{\code{method}}{\link[R6:R6Class]{\code{method}}}}
argument of
\ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-declare_initialization"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-declare_initialization}{}}}
\subsection{Method \code{declare_initialization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$declare_initialization()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Create (initial) assignment \ifelse{html}{\link[base:expression]{\code{expression}}}{\ifelse{latex}{\code{expression}}{\link[base:expression]{\code{expression}}}} for
\code{R6Member} object.
\if{html}{\out{<br>}}\emph{To be used when declaring \ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}},
see also
the
\ifelse{html}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{R6MemberList$R6Class()}}}{\link[=R6MemberList]{\code{R6MemberList$R6Class()}}}}
method
and
\code{?R6Class}.
}

\subsection{Returns}{
\ifelse{html}{\link[base:expression]{\code{expression}}}{\ifelse{latex}{\code{expression}}{\link[base:expression]{\code{expression}}}} containing the assignment statement
to be used in the
\ifelse{html}{\link[R6:R6Class]{\code{public[["initialize"]]}}}{\ifelse{latex}{\code{public[["initialize"]]}}{\link[R6:R6Class]{\code{public[["initialize"]]}}}}
(constructor)
method of
\ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6Member-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6Member-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6Member$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
