% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-R6MemberList.R
\name{R6MemberList}
\alias{R6MemberList}
\title{\code{R6MemberList}:
Wrapping toolbox for using
\code{R6Members}
in \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}} definitions.}
\description{
A \code{R6MemberList} provides utility
\ifelse{html}{\href{#method-R6MemberList-active_bindings}{methods}}{\ifelse{latex}{\href{#method-R6MemberList-active_bindings}{methods}}{\link[=R6MemberList]{methods}}}
for declaring
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}} with members wrapped in
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects.
As such, the \code{R6MemberList} encloses zero or more
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects that
define the members in
a \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}}.
\if{html}{\out{<br>}}
The core method for this purpose is \ifelse{html}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\link[=R6MemberList]{\code{$R6Class()}}}}.
}
\section{Details}{

Once a \code{R6MemberList} is instanciated,
the
\ifelse{html}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\link[=R6MemberList]{\code{$R6Class()}}}}
method can be used
to define a \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}}
and delegate the enclosed
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects
as members to this new \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}}.

The core method for this purpose is \ifelse{html}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\link[=R6MemberList]{\code{$R6Class()}}}},
which calls the worker methods
\ifelse{html}{\href{#method-R6MemberList-public_list}{\code{$public_list()}}}{\ifelse{latex}{\href{#method-R6MemberList-public_list}{\code{$public_list()}}}{\link[=R6MemberList]{\code{$public_list()}}}},  \ifelse{html}{\href{#method-R6MemberList-private_list}{\code{$private_list()}}}{\ifelse{latex}{\href{#method-R6MemberList-private_list}{\code{$private_list()}}}{\link[=R6MemberList]{\code{$private_list()}}}} and \ifelse{html}{\href{#method-R6MemberList-active_bindings}{\code{$active_bindings()}}}{\ifelse{latex}{\href{#method-R6MemberList-active_bindings}{\code{$active_bindings()}}}{\link[=R6MemberList]{\code{$active_bindings()}}}}.
These set up the \ifelse{html}{\link[base:list]{\code{lists}}}{\ifelse{latex}{\code{lists}}{\link[base:list]{\code{lists}}}} defining the arguments
\ifelse{html}{\link[R6:R6Class]{\code{public}}}{\ifelse{latex}{\code{public}}{\link[R6:R6Class]{\code{public}}}},  \ifelse{html}{\link[R6:R6Class]{\code{private}}}{\ifelse{latex}{\code{private}}{\link[R6:R6Class]{\code{private}}}} and \ifelse{html}{\link[R6:R6Class]{\code{active}}}{\ifelse{latex}{\code{active}}{\link[R6:R6Class]{\code{active}}}}
for the call to \ifelse{html}{\link[R6:R6Class]{\code{R6::R6Class()}}}{\ifelse{latex}{\code{R6::R6Class()}}{\link[R6:R6Class]{\code{R6::R6Class()}}}}.
\subsection{Retrieving private and public names}{

Use \ifelse{html}{\href{#method-R6MemberList-names}{\code{$names()}}}{\ifelse{latex}{\href{#method-R6MemberList-names}{\code{$names()}}}{\link[=R6MemberList]{\code{$names()}}}} to retrieve public or private member names
of the \ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects in the \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}}
created by \ifelse{html}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\ifelse{latex}{\href{#method-R6MemberList-R6Class}{\code{$R6Class()}}}{\link[=R6MemberList]{\code{$R6Class()}}}}.
}

\subsection{Subsetting}{

The
\ifelse{html}{\href{#method-R6MemberList-select}{\code{$select()}}}{\ifelse{latex}{\href{#method-R6MemberList-select}{\code{$select()}}}{\link[=R6MemberList]{\code{$select()}}}} method
allows access to the enclosed \ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects
by name.
Alternatively, the \ifelse{html}{\link[utils:methods]{\code{S3-operators}}}{\ifelse{latex}{\code{S3-operators}}{\link[utils:methods]{\code{S3-operators}}}}
\ifelse{html}{\link[base:[]{\code{[}}}{\ifelse{latex}{\code{[}}{\link[base:[]{\code{[}}}} and \ifelse{html}{\link[base:[[]{\code{[[}}}{\ifelse{latex}{\code{[[}}{\link[base:[[]{\code{[[}}}}
allow subsetting for \code{R6MemberList} as well.

See the
\ifelse{html}{\href{#method-R6MemberList-new}{methods}}{\ifelse{latex}{\href{#method-R6MemberList-new}{methods}}{\link[=R6MemberList]{methods}}} and \ifelse{html}{\href{#sec-R6MemberList-examples}{examples}}{\ifelse{latex}{\href{#sec-R6MemberList-examples}{examples}}{\link[=R6MemberList]{examples}}}
section for more details and examples.
}
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# ===================== > Create R6MemberRestrictions < ====================== #

# Restriction that checks if length of value equals 2
length_restriction <-
    R6MemberRestriction$new(
        id = "Length check",
        restriction = function(x) \{
            if (length(x) != 2) \{
                stop("Length should be 2!")
            \}
            return(TRUE)
        \}
    )

# Restriction that checks if type of value is "character"
type_restriction <-
    R6MemberRestriction$new(
        id = "Type check",
        restriction = function(x) \{
            if (typeof(x) != "character") \{
                stop("type should be character!")
            \}
            return(TRUE)
        \}
    )
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# =========================== > Create R6Members < =========================== #

# R6Member object without restriction(s):
#   may take any value
R6Member_any <-
   R6Member$new(
       name = "example_member_any",
       value = 1:5
   )

# R6Member object with restriction(s):
#   all restriction(s) check(s) must return TRUE
R6Member_two_string <-
   R6Member$new(
       name = "example_member_char_len2",
       value = c("foo", "bar"),
       restrictions = list(
               "is length 2" = length_restriction,
               "is character" = type_restriction
       )
   )
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# ========================= > Create R6MemberList < ========================== #

R6MemberList_example <-
   R6MemberList$new(
       # R6Member object without restrictions
       R6Member_any,
       # R6Member object with restriction(s)
       R6Member_two_string
   )
}\if{html}{\out{</div>}}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6MemberList-new}{\code{R6MemberList$new()}}
\item \href{#method-R6MemberList-combine}{\code{R6MemberList$combine()}}
\item \href{#method-R6MemberList-active_bindings}{\code{R6MemberList$active_bindings()}}
\item \href{#method-R6MemberList-R6Class}{\code{R6MemberList$R6Class()}}
\item \href{#method-R6MemberList-public_list}{\code{R6MemberList$public_list()}}
\item \href{#method-R6MemberList-private_list}{\code{R6MemberList$private_list()}}
\item \href{#method-R6MemberList-print}{\code{R6MemberList$print()}}
\item \href{#method-R6MemberList-names}{\code{R6MemberList$names()}}
\item \href{#method-R6MemberList-select}{\code{R6MemberList$select()}}
\item \href{#method-R6MemberList-clone}{\code{R6MemberList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{R6MemberList} object
from the supplied \ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}}
objects to be combined in the
\code{R6MemberList}
\if{html}{\out{<br>}}\emph{Objects of other types are not accepted.}\tabular{ll}{
    \tab  \cr
   Type:   \tab \ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{R6MemberList} object containing the provided
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-combine"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-combine}{}}}
\subsection{Method \code{combine()}}{
Combine with another \code{R6MemberList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$combine(obj)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{R6MemberList}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-active_bindings"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-active_bindings}{}}}
\subsection{Method \code{active_bindings()}}{
Create named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the active bindings provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{active}'}}{\ifelse{latex}{'\code{active}'}{\link[R6:R6Class]{'\code{active}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition. \if{html}{\out{<br>}}\emph{See also the documentation of argument \ifelse{html}{\link[R6:R6Class]{'\code{active}'}}{\ifelse{latex}{'\code{active}'}{\link[R6:R6Class]{'\code{active}'}}} in \code{?R6Class} .}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$active_bindings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the active bindings provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{active}'}}{\ifelse{latex}{'\code{active}'}{\link[R6:R6Class]{'\code{active}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-R6Class"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-R6Class}{}}}
\subsection{Method \code{R6Class()}}{
Declare \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}} from \code{R6MemberList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$R6Class(classname, print_method = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classname}}{Name of the
\ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}}
to be defined.
See also the documentation of argument
\ifelse{html}{\link[R6:R6Class]{'\code{classname}'}}{\ifelse{latex}{'\code{classname}'}{\link[R6:R6Class]{'\code{classname}'}}}
in
\code{?R6Class}.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{print_method}}{\if{html}{\out{
<!-- ```\{r,child = "../examples/class-R6MemberList-initialize.Rmd",eval=TRUE,echo=TRUE\}
``` -->
}}


The print method to use for the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
to be defined.
May be one of the following:\tabular{ll}{
    \tab  \cr
   \code{"none"} or \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \tab No own print method is specified. The parent class' print method is inherited. \cr
   \code{"default"}              \tab A default print method is used.  \cr
   \ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}} \tab The provided function will be used as print method. \cr
}
\tabular{ll}{
    \tab  \cr
   Type:   \tab \ifelse{html}{\link[base:character]{\code{character}}}{\ifelse{latex}{\code{character}}{\link[base:character]{\code{character}}}},  \ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}} or  \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} (See above.) \cr
   Default:  \tab \code{"default"}        \cr
}
\if{html}{\out{
<br>
}}


\strong{The \code{"default"} print method}
\if{html}{\out{<br>}}
takes the same arguments as\ifelse{html}{\link[=R6Member]{\code{R6Member$print()}}}{\ifelse{latex}{\href{#method-R6Member-print}{\code{R6Member$print()}}}{\link[=R6Member]{\code{R6Member$print()}}}}, namely \code{value_level},  \code{restriction_level} and \code{print_class} .
\if{html}{\out{<br>}}
It generates output like in the following example:
\if{html}{\out{<br>}}\preformatted{

Object of class(es)
   `"ExampleR6Class"`
Members:
R6Member example_member_any
   Value:
      [1] 1 2 3 4 5 
R6Member example_member_char_len2
   Value:
      [1] "foo" "bar" 

}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \ifelse{html}{\link[R6:R6Class]{\code{R6Class}}}{\ifelse{latex}{\code{R6Class}}{\link[R6:R6Class]{\code{R6Class}}}} with members defined by the \code{R6MemberList}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-public_list"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-public_list}{}}}
\subsection{Method \code{public_list()}}{
Create named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the public fields provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{public}'}}{\ifelse{latex}{'\code{public}'}{\link[R6:R6Class]{'\code{public}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition. \if{html}{\out{<br>}}\emph{See also the documentation of argument \ifelse{html}{\link[R6:R6Class]{'\code{public}'}}{\ifelse{latex}{'\code{public}'}{\link[R6:R6Class]{'\code{public}'}}} in \code{?R6Class} .}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$public_list(print_method = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print_method}}{\if{html}{\out{
<!-- ```\{r,child = "../examples/class-R6MemberList-initialize.Rmd",eval=TRUE,echo=TRUE\}
``` -->
}}


The print method to use for the
\ifelse{html}{\link[R6:R6Class]{\code{R6Classes}}}{\ifelse{latex}{\code{R6Classes}}{\link[R6:R6Class]{\code{R6Classes}}}}
to be defined.
May be one of the following:\tabular{ll}{
    \tab  \cr
   \code{"none"} or \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} \tab No own print method is specified. The parent class' print method is inherited. \cr
   \code{"default"}              \tab A default print method is used.  \cr
   \ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}} \tab The provided function will be used as print method. \cr
}
\tabular{ll}{
    \tab  \cr
   Type:   \tab \ifelse{html}{\link[base:character]{\code{character}}}{\ifelse{latex}{\code{character}}{\link[base:character]{\code{character}}}},  \ifelse{html}{\link[base:function]{\code{function}}}{\ifelse{latex}{\code{function}}{\link[base:function]{\code{function}}}} or  \ifelse{html}{\link[base:NULL]{\code{NULL}}}{\ifelse{latex}{\code{NULL}}{\link[base:NULL]{\code{NULL}}}} (See above.) \cr
   Default:  \tab \code{"default"}        \cr
}
\if{html}{\out{
<br>
}}


\strong{The \code{"default"} print method}
\if{html}{\out{<br>}}
takes the same arguments as\ifelse{html}{\link[=R6Member]{\code{R6Member$print()}}}{\ifelse{latex}{\href{#method-R6Member-print}{\code{R6Member$print()}}}{\link[=R6Member]{\code{R6Member$print()}}}}, namely \code{value_level},  \code{restriction_level} and \code{print_class} .
\if{html}{\out{<br>}}
It generates output like in the following example:
\if{html}{\out{<br>}}\preformatted{

Object of class(es)
   `"ExampleR6Class"`
Members:
R6Member example_member_any
   Value:
      [1] 1 2 3 4 5 
R6Member example_member_char_len2
   Value:
      [1] "foo" "bar" 

}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the public fields provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{public}'}}{\ifelse{latex}{'\code{public}'}{\link[R6:R6Class]{'\code{public}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-private_list"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-private_list}{}}}
\subsection{Method \code{private_list()}}{
Create named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the private fields provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{private}'}}{\ifelse{latex}{'\code{private}'}{\link[R6:R6Class]{'\code{private}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition. \if{html}{\out{<br>}}\emph{See also the documentation of argument \ifelse{html}{\link[R6:R6Class]{'\code{private}'}}{\ifelse{latex}{'\code{private}'}{\link[R6:R6Class]{'\code{private}'}}} in \code{?R6Class} .}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$private_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named \ifelse{html}{\link[base:list]{\code{list}}}{\ifelse{latex}{\code{list}}{\link[base:list]{\code{list}}}} that represent the private fields provided by the stored \ifelse{html}{\link[=R6Member]{\code{R6Members}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Members}}}{\link[=R6Member]{\code{R6Members}}}}. \if{html}{\out{<br>}} To be used as argument \ifelse{html}{\link[R6:R6Class]{'\code{private}'}}{\ifelse{latex}{'\code{private}'}{\link[R6:R6Class]{'\code{private}'}}} in \ifelse{html}{\link[R6:R6Class]{\code{R6Class()}}}{\ifelse{latex}{\code{R6Class()}}{\link[R6:R6Class]{\code{R6Class()}}}} definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-print"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{R6MemberList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-names}{}}}
\subsection{Method \code{names()}}{
\code{names} method for \code{R6MemberList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$names(type = "public", subset = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of names to be returned.
One of \code{c("public", "private")}.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
   Default:  \tab \code{"public"}         \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{subset}}{The subset of
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}}
names from the
\code{R6MemberList}
to be returned.
One of \code{c("all", "public", "private")}.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
   Default:  \tab \code{"all"}         \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-select"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-select}{}}}
\subsection{Method \code{select()}}{
Subsetting method for \code{R6MemberList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$select(member, new_value, inner = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{member}}{Names of the
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}}
objects in the
\code{R6MemberList}
to return.\tabular{ll}{
    \tab  \cr
   Type:   \tab character          \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{new_value}}{Either missing (for the getter),
or the new value(s) to be assigned to the
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}}
objects in the
\code{R6MemberList}
(setter case).\tabular{ll}{
    \tab  \cr
   Type:   \tab any           \cr
}
\if{html}{\out{
 <br>
}}}

\item{\code{inner}}{If \code{TRUE}, the subset is a \emph{single}
\ifelse{html}{\link[=R6Member]{\code{R6Member}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member}}}{\link[=R6Member]{\code{R6Member}}}}
object from the
\code{R6MemberList}, in analogy to \code{list[[elem]]}.\if{html}{\out{<br>}}
If \code{FALSE}, the subset itself is a
\code{R6MemberList}
containing
\ifelse{html}{\link[=R6Member]{\code{R6Member objects}}}{\ifelse{latex}{\href{#R6Member}{\code{R6Member objects}}}{\link[=R6Member]{\code{R6Member objects}}}}, in analogy to \code{list[elems]}.\tabular{ll}{
    \tab  \cr
   Type:   \tab logical          \cr
   Default:  \tab \code{TRUE}         \cr
}
\if{html}{\out{
 <br>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
select
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MemberList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6MemberList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MemberList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
